Move publishing under sub-package
